backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
httpx>=0.27.2
python-dotenv>=1.0.1
groq>=0.11.0
supabase==2.7.4
pydantic>=2.10.0
